Moved pulse ui tests into their own package.
